+++ server/routes/propertyRoutes.js
@@ -11,7 +11,7 @@ router.get('/properties', PropertyController.viewAllProperties);
 router.get('/property/:id', PropertyController.viewPropertyById);
 router.post('/property', multipartyMiddle, ValidationMiddleware.createPropertyValidation, PropertyController.postNewProperty);
 router.delete('/property/:id', PropertyController.deleteProperty);
-router.patch('/property/:id', multipartyMiddle, PropertyController.updatePropertyDetails);
+router.patch('/property/:id', multipartyMiddle, ValidationMiddleware.updatePropertyValidation, PropertyController.updatePropertyDetails);
 router.patch('/property/:id/sold', PropertyController.markAsSold);
 router.get('/property', PropertyController.viewPropertiesByType);
